# contracts/control-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: AstrLink Local Control API
  version: 1.0.0
  summary: Stable local contract between the AstrLink desktop application and Core.
  description: |
    This API is a local control plane; it is not part of the public inference API.
    Core binds the inference plane to `127.0.0.1:8317` by default and binds this
    control plane to the ephemeral address `127.0.0.1:0`. On startup Core writes
    exactly one ready JSON line to stdout containing the selected `control_url`.
    Launchers must use that value rather than guessing a port.

    The bootstrap health, version, and capability operations are intentionally
    unauthenticated and disclose no configuration. Every operation that reads or
    mutates user configuration, events, or request records requires the per-start
    local control token. The token must be delivered through an inherited pipe or
    equivalent protected IPC; it must not be placed in process arguments or the
    environment.

    Request and response bodies are excluded from ordinary request records. The two
    audit-body switches are independent, default to false, and can only be enabled
    through this authenticated control API after explicit risk acknowledgement.

    The three bootstrap GET operations are always available. Authenticated Service
    CRUD, service-scoped subscription OAuth, priority Route CRUD, persistent local
    access-token operations, the fixed
...
    capability claim.
servers:
...
    variables:
...
security:
  - controlToken: []
...
paths:
...
  /control/v1/routes:
    x-astrlink-implementation-status: implemented
    get:
      tags: [Routes]
      operationId: listRoutes
      summary: List routes in deterministic evaluation order.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: enabled
          in: query
          schema:
            type: boolean
      deprecated: true
      responses:
        '410':
          description: Legacy routing feature retired; use service order and routing settings
        "200":
          description: Route page.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Routes]
      operationId: createRoute
      summary: Create a route.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteCreate"
      deprecated: true
      responses:
        '410':
          description: Legacy routing feature retired; use service order and routing settings
        "201":
          description: Route created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /control/v1/routes/{route_id}:
    x-astrlink-implementation-status: implemented
    parameters:
      - $ref: "#/components/parameters/RouteId"
    get:
      tags: [Routes]
      operationId: getRoute
      summary: Read one route.
      deprecated: true
      responses:
        '410':
          description: Legacy routing feature retired; use service order and routing settings
        "200":
          description: Route details.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Routes]
      operationId: updateRoute
      summary: Update a route.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/RoutePatch"
      deprecated: true
      responses:
        '410':
          description: Legacy routing feature retired; use service order and routing settings
        "200":
          description: Updated route.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      tags: [Routes]
      operationId: deleteRoute
      summary: Delete a route.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      deprecated: true
      responses:
        '410':
          description: Legacy routing feature retired; use service order and routing settings
        "204":
          description: Route deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
...
components:
  securitySchemes:
    controlToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-control-token
      description: Per-start local token delivered over protected IPC; never persist it in configuration or logs.
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the previous page. It must not be parsed by clients.
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
    RouteId:
      name: route_id
      in: path
      required: true
      schema:
        $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag from the latest GET; prevents lost updates.
      schema:
        type: string
        minLength: 3
        maxLength: 128
  headers:
    NoStore:
      description: Sensitive local-control responses must not be cached.
      schema:
        type: string
        examples: [no-store]
    Location:
      description: Relative path of the created resource.
      schema:
        type: string
        format: uri-reference
    ETag:
      description: Strong entity tag for the current resource representation.
      schema:
        type: string
  responses:
    BadRequest:
      description: Malformed request, query, or cursor.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid local control token.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The operation conflicts with current state or references.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PreconditionFailed:
      description: If-Match does not match the current resource representation.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: The request is well-formed but violates a contract or safety invariant.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    FailurePolicy:
      type: object
      required: [max_retries, initial_delay_ms, max_delay_ms, network_error, response_timeout, http_status]
      properties:
        max_retries:
          type: integer
          minimum: 0
          maximum: 5
          default: 1
        initial_delay_ms:
          type: integer
          minimum: 0
          maximum: 60000
          default: 500
        max_delay_ms:
          type: integer
          minimum: 0
          maximum: 60000
          default: 5000
          description: Must be at least initial_delay_ms. A longer Retry-After prevents retrying this target and permits switching only if the error rule allows it.
        response_start_timeout_seconds:
          type: integer
          minimum: 0
          maximum: 86400
          description: Omission uses the global timeout. Zero is unlimited. Never imposes a deadline on a started stream.
        thinking_signature_recovery:
          type: boolean
          default: true
          description: On a recognized Claude Messages thinking signature HTTP 400, repair assistant history and retry the same target once within existing attempt limits. Explicit HTTP 400 rules prohibiting retries take precedence. Manual paths require a later step for the same target. Omission enables recovery; false disables it.
        openai_reasoning_recovery:
          type: boolean
          default: true
          description: On a recognized OpenAI Responses or Responses Compact reasoning HTTP 400, repair rejected reasoning ciphertext or identified item fields and retry the same target once within existing attempt limits. Preserves summaries, tool pairs, previous_response_id and compaction context. Explicit HTTP 400 rules prohibiting retries take precedence. Manual paths require a later step for the same target. Omission enables recovery; false disables it.
        openai_function_output_recovery:
          type: boolean
          default: false
          description: Opt-in. On a recognized OpenAI Responses or Responses Compact HTTP 400 for encrypted function or custom-tool output, drop encrypted output parts, keep plaintext, and retry the same target once within existing attempt limits. When a reasoning ciphertext repair already runs, this also clears those output parts in the same attempt. Tool call_id pairs and compaction stay intact. Explicit HTTP 400 rules prohibiting retries take precedence. Omission disables recovery; true enables it.
        network_error:
          $ref: "#/components/schemas/FailureAction"
        response_timeout:
          $ref: "#/components/schemas/FailureAction"
        http_status:
          type: object
          propertyNames:
            pattern: '^[45][0-9]{2}$'
          additionalProperties:
            $ref: "#/components/schemas/FailureAction"
          description: Unlisted statuses stop, except recognized thinking signature, OpenAI reasoning, or opt-in encrypted function-output errors when the corresponding repair switch is enabled. Defaults permit retry and switching for 408, 429, 500, 502, 503, 504, 529; 401 and 403 permit switching only.
      additionalProperties: false
      description: A complete policy. Missing service policy inherits RoutingSettings.default_failure_policy; missing route policy uses each service policy (or the global policy). A route override replaces the entire inherited policy. Missing route failover inherits global strategy and max_attempts with switching enabled.
    FailoverPolicy:
      type: object
      required: [enabled, strategy, max_attempts]
      properties:
        enabled:
          type: boolean
          default: true
        strategy:
          type: string
          enum: [retry_first, failover_first, failover_only]
          default: retry_first
        max_attempts:
          type: integer
          minimum: 1
          maximum: 20
          default: 6
          description: Total actual upstream inference requests, including the first request, retries and backups.
      additionalProperties: false
...
    AstrLinkAutoModel:
      type: string
      const: astrlink/auto
      description: >
        Reserved public virtual-model identifier. It is never an ordinary
        alias: requests using this exact model require a Route whose
        selection.mode is auto.
...
    RouteSelection:
      type: object
      required: [mode]
      properties:
        mode:
          $ref: "#/components/schemas/RouteSelectionMode"
        taxonomy_id:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-z][a-z0-9._-]{0,63}$'
          example: astrlink-text-v1
          description: >
            Versioned task-classification taxonomy used by auto selection.
            Required when mode is auto and forbidden when mode is priority.
            It identifies label semantics, not a classifier implementation.
      allOf:
        - if:
            properties:
              mode:
                const: auto
            required: [mode]
          then:
            required: [taxonomy_id]
          else:
            not:
              required: [taxonomy_id]
      additionalProperties: false
      description: >
        Optional target-selection strategy. Omission is exactly equivalent to
        mode priority, preserving existing Route behavior. Mode auto requires
        match.model to equal astrlink/auto and replaces the ordinary top-level
        targets array with at least two RouteCategory objects. Every category
        owns one or more targets, every target has a concrete upstream_model,
        and the complete Route contains at least two distinct upstream model
        names. The taxonomy ID and category-owned model pools are stable
        configuration; classifier weights, confidence values, and model-ranking
        predictions are not.
    Route:
      type: object
      required:
        - id
        - name
        - enabled
        - priority
        - match
      properties:
        recovery_path_id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        failure_policy:
          $ref: "#/components/schemas/FailurePolicy"
        failover:
          $ref: "#/components/schemas/FailoverPolicy"
        id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        name:
          type: string
          minLength: 1
          maxLength: 128
        enabled:
          type: boolean
        priority:
          type: integer
          minimum: 0
          maximum: 1000000
        match:
          $ref: "#/components/schemas/RouteMatch"
        selection:
          $ref: "#/components/schemas/RouteSelection"
        targets:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RouteTarget"
        categories:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/RouteCategory"
      allOf:
        - if:
            properties:
              selection:
                properties:
                  mode:
                    const: auto
                required: [mode]
            required: [selection]
          then:
            required: [categories]
            properties:
              match:
                properties:
                  model:
                    $ref: "#/components/schemas/AstrLinkAutoModel"
                required: [model]
            not:
              anyOf:
                - required: [targets]
                - required: [recovery_path_id]
          else:
            oneOf:
              - required: [targets]
              - required: [recovery_path_id]
            properties:
              match:
                not:
                  properties:
                    model:
                      $ref: "#/components/schemas/AstrLinkAutoModel"
                  required: [model]
            not:
              required: [categories]
      additionalProperties: false
    RouteCreate:
      type: object
      required: [name, priority, match]
      properties:
        recovery_path_id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        failure_policy:
          $ref: "#/components/schemas/FailurePolicy"
        failover:
          $ref: "#/components/schemas/FailoverPolicy"
        name:
          type: string
          minLength: 1
          maxLength: 128
        enabled:
          type: boolean
          default: true
        priority:
          type: integer
          minimum: 0
          maximum: 1000000
        match:
          $ref: "#/components/schemas/RouteMatch"
        selection:
          $ref: "#/components/schemas/RouteSelection"
        targets:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RouteTarget"
        categories:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/RouteCategory"
      allOf:
        - if:
            properties:
              selection:
                properties:
                  mode:
                    const: auto
                required: [mode]
            required: [selection]
          then:
            required: [categories]
            properties:
              match:
                properties:
                  model:
                    $ref: "#/components/schemas/AstrLinkAutoModel"
                required: [model]
            not:
              anyOf:
                - required: [targets]
                - required: [recovery_path_id]
          else:
            oneOf:
              - required: [targets]
              - required: [recovery_path_id]
            properties:
              match:
                not:
                  properties:
                    model:
                      $ref: "#/components/schemas/AstrLinkAutoModel"
                  required: [model]
            not:
              required: [categories]
      additionalProperties: false
    RoutePatch:
      type: object
      minProperties: 1
      properties:
        recovery_path_id:
          oneOf:
            - $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
            - type: "null"
        failure_policy:
          oneOf:
            - $ref: "#/components/schemas/FailurePolicy"
            - type: "null"
        failover:
          oneOf:
            - $ref: "#/components/schemas/FailoverPolicy"
            - type: "null"
        name:
          type: string
          minLength: 1
          maxLength: 128
        enabled:
          type: boolean
        priority:
          type: integer
          minimum: 0
          maximum: 1000000
        match:
          $ref: "#/components/schemas/RouteMatch"
        selection:
          oneOf:
            - $ref: "#/components/schemas/RouteSelection"
            - type: "null"
              description: Restore the backward-compatible priority selection mode.
        targets:
          oneOf:
            - type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/RouteTarget"
            - type: "null"
              description: Remove top-level priority targets when switching to auto.
        categories:
          oneOf:
            - type: array
              minItems: 2
              items:
                $ref: "#/components/schemas/RouteCategory"
            - type: "null"
              description: Remove category pools when switching to priority.
          description: >
            After applying a merge patch, the resulting Route must satisfy the
            priority/auto conditional shape and full cross-field validation.
      additionalProperties: false
    RouteMatch:
      type: object
      required: [protocol]
      properties:
        protocol:
          $ref: ./protocol-capabilities.schema.json#/$defs/ProtocolId
        model:
          type: string
          minLength: 1
          maxLength: 256
          description: >
            Exact public model or alias. Omission matches any model.
            astrlink/auto is reserved and is valid only with
            selection.mode=auto.
      additionalProperties: false
    RouteTarget:
      type: object
      required:
        - service_id
        - plan_type
        - upstream_protocol
        - priority
      properties:
        service_id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        plan_type:
          $ref: ./protocol-capabilities.schema.json#/$defs/PlanType
          description: RelayKit is a stable future value but is rejected while capabilities report it unavailable in Alpha.
        upstream_protocol:
          $ref: ./protocol-capabilities.schema.json#/$defs/ProtocolId
        priority:
          type: integer
          minimum: 0
          maximum: 1000000
        upstream_model:
          type: string
          minLength: 1
          maxLength: 256
          description: >
            Rewrite the public model to this upstream model for this target.
            Requires a non-empty match.model on the route. Omission preserves
            the public model bytes. Every auto-selection target requires this
            field; target priority remains its deterministic low-confidence and
            transport-failure fallback order.
      additionalProperties: false
...
    RoutePage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Route"
        next_cursor:
          type: [string, "null"]
          maxLength: 512
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/ControlError"
        request_id:
          type: string
          minLength: 1
          maxLength: 96
      additionalProperties: false
...